Rename json_operations to json
