+++ crates/parameters/src/lib.rs
@@ -1,2 +1,2 @@
 pub mod directory;
-pub mod json_operations;
+pub mod json;
